pyproject.toml
[tool.poetry]
name = "taskmanagerapi"
version = "0.1.0"
description = ""
authors = ["Hordunlarmy <Hordunlarmy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"
python-decouple = "^3.8"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.4.0"
drf-standardized-errors = "^0.14.1"
django-cors-headers = "^4.6.0"
django-debug-toolbar = "^5.0.1"
psycopg2 = "^2.9.10"
uvicorn = "^0.34.0"
drf-spectacular = {extras = ["swagger-ui"], version = "^0.28.0"}
django-rest-swagger = "^2.2.0"
drf-yasg = "^1.21.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
